--- mahasiswa/dashboard.pug
@@ -1,3 +1,0 @@
-extends layout/master
-
-block web-content
