--- master.pug
@@ -1,88 +1,0 @@
-doctype html
-html(lang="en")
-    head
-        meta(charset="utf-8")
-        meta(http-equiv="X-UA-Compatible", content="IE=edge")
-        meta(name="viewport", content="width=device-width, initial-scale=1, shrink-to-fit=no")
-        meta(name="description", content="")
-        meta(name="author", content="")
-
-        block title
-
-        // Custom fonts for this template
-        link(href="/vendor/fontawesome-free/css/all.min.css", rel="stylesheet", type="text/css")
-        link(href="https://fonts.googleapis.com/css?family=Nunito:200,200i,300,300i,400,400i,600,600i,700,700i,800,800i,900,900i", rel="stylesheet")
-
-        // Custom styles for this template
-        link(href="/css/sb-admin-2.min.css", rel="stylesheet")
-        block custom-css
-
-    body(id="page-top")
-        // Page Wrapper
-        #wrapper
-            // Sidebar
-            block sidebar
-                // End of Sidebar
-
-            // Content Wrapper
-            #content-wrapper.d-flex.flex-column
-                // Main Content
-                #content
-                    // Topbar
-                    block top-bar
-                        // Default content for top-bar
-                        nav.navbar.navbar-expand.navbar-light.bg-white.topbar.mb-4.static-top.shadow
-                            button#sidebarToggleTop.btn.btn-link.d-md-none.rounded-circle.mr-3
-                                i.fa.fa-bars
-
-                    // Begin Page Content
-                    .container
-                        block web-content
-
-                    // End of Main Content
-
-                    // Footer
-                    footer.sticky-footer.bg-white
-                        .container.my-auto
-                            .copyright.text-center.my-auto
-                                span Copyright &copy; Your Website 2021
-
-                    // End of Footer
-
-                // End of Content Wrapper
-
-            // End of Page Wrapper
-
-            // Scroll to Top Button
-            a.scroll-to-top.rounded(href="#page-top")
-                i.fas.fa-angle-up
-
-            // Logout Modal
-            .modal.fade#logoutModal(tabindex="-1", role="dialog", aria-labelledby="exampleModalLabel", aria-hidden="true")
-                .modal-dialog(role="document")
-                    .modal-content
-                        .modal-header
-                            h5.modal-title#exampleModalLabel Ready to Leave?
-                            button.close(type="button", data-dismiss="modal", aria-label="Close")
-                                span(aria-hidden="true") ×
-                        .modal-body Select "Logout" below if you are ready to end your current session.
-                        .modal-footer
-                            button.btn.btn-secondary(type="button", data-dismiss="modal") Cancel
-                            a.btn.btn-primary(href="login.pug") Logout
-
-        // Bootstrap core JavaScript
-        script(src="/vendor/jquery/jquery.min.js")
-        script(src="/vendor/bootstrap/js/bootstrap.bundle.min.js")
-
-        // Core plugin JavaScript
-        script(src="/vendor/jquery-easing/jquery.easing.min.js")
-
-        // Custom scripts for all pages
-        script(src="/js/sb-admin-2.min.js")
-
-        // Page level plugins
-        script(src="/vendor/chart.js/Chart.min.js")
-
-        // Page level custom scripts
-        script(src="/js/demo/chart-area-demo.js")
-        script(src="/js/demo/chart-pie-demo.js")
